lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
pytz
ask-sdk-dynamodb-persistence-adapter==1.15.0